+++ rollup.config.js
@@ -1,4 +1,6 @@
 import cjs from '@rollup/plugin-commonjs'
+import nodePolyfills from 'rollup-plugin-node-polyfills';
+import { nodeResolve } from '@rollup/plugin-node-resolve';
 import {dependencies} from './package.json'
 
 const plugins = [cjs()]
@@ -10,11 +12,22 @@ export default [
     input: 'src/index.js',
     output: [
       {
-        file: 'dist/index.js',
+        file: 'dist/index.cjs',
         format: 'cjs',
       },
     ],
     external,
     plugins,
   },
+  {
+    input: 'src/browser.js',
+    output: [
+        {
+            file: 'dist/browser.es.js',
+            format: 'esm',
+          },
+    ],
+    plugins: [nodeResolve({browser: true}), ...plugins,nodePolyfills()],
+  },
 ]
+