+++ changelog.md
@@ -3,7 +3,8 @@
 ## 3.0.0
 
  * Remove filename option from parse
-
+ * Works in the browser
+ * Rename dist/index.js dist/index.cjs
 
 ## 2.1.0